refactor: remove unneeded children from ClickedObject
