+++ src/shared/util.ts
@@ -415,44 +415,3 @@ export function normalizePath(p: string): string {
 
   return p
 }
-
-export function findSubTree(tree: GitTreeObject, path?: string): GitTreeObject {
-  if (!path) return tree
-
-  const hasRootPrefix = path === tree.name || path.startsWith(`${tree.name}/`)
-  if (!hasRootPrefix) return tree
-
-  let relativePath = path.length === tree.name.length ? "" : path.substring(tree.name.length + 1)
-  if (!relativePath) return tree
-
-  relativePath = normalizePath(relativePath)
-
-  let currentTree: GitTreeObject = tree
-  const separator = getSep(relativePath)
-  const steps = relativePath.split(separator)
-
-  for (let i = 0; i < steps.length; i++) {
-    let foundStep = false
-
-    for (const child of currentTree.children) {
-      if (child.type === "tree") {
-        const childSteps = child.name.split(separator)
-
-        if (childSteps[0] === steps[i]) {
-          currentTree = child
-          i += childSteps.length - 1
-          foundStep = true
-          break
-        }
-      }
-    }
-
-    // If a step is not found, abort and return root tree
-    if (!foundStep) {
-      console.warn(`Could not find step ${steps[i]} in subtree ${currentTree.name}`)
-      return tree
-    }
-  }
-
-  return currentTree
-}
